Create lib.rs to enable running cranks as libs
